requirements.txt
flask
langchain_text_splitters
python-dotenv
pydantic
PyPDF2
python-docx
chromadb
ollama
docling